requirements.txt
pandas~=1.4.2
GitPython~=3.1.27
pycg
astor~=0.8.1
graphviz~=0.17
networkx
astunparse~=1.6.3
typed-ast
typeshed-client
regex
setuptools
dataclasses
networkx
packaging